+++ src/Middleware/Resource/Resources.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Flipbox\Relay\Salesforce\Middleware\Resource;
+
+use Flipbox\Relay\Salesforce\Middleware\Resource;
+
+class Resources extends Resource
+{
+    /**
+     * The resource name
+     */
+    public $resource = '';
+}