+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All Notable changes to `flipboxdigital\relay-salesforce` will be documented in this file
 
 ## Unreleased
+### Added
+- [Resources by Version](https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/resources_discoveryresource.htm) middleware and pipeline builder
+
 ### Changed
 - Fixed incorrect describe global resource endpoint